+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # next [????.??.??]
+* Backport the `Generic(1)` instances for `Data.Ord.Down`, introduced in
+  `base-4.12`.
 * Refactor internals using `th-abstraction`.
 * Adapt to `Maybe` moving to `GHC.Maybe` in GHC 8.6.